Refactor tests for proper collection
